requirements.txt
nextcord==2.0.0b4
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.28.0
